ENH: Aggregate Expert
This expert is able to aggregate events with a defined threshold.

How does it work?
The expert will use specified fields from an event as identifier.
Then stores how often the event occur and if the threshold has
been reached within its set timespan, it will push the event to
the queue, else it will just get dropped.

Signed-off-by: Sebastian Waldbauer <waldbauer@cert.at>